Add DependencyGraph.contains (#8)
Methods that check for membership of vertices are required for more sophisticated methods. Examples for such methods are adding a vertex or finding all neighbours of a vertex. When adding a vertex, we first want to make sure that the vertex does not exist already because we do not want to override its entries in the dictionaries.

TestGraph was added to avoid code duplication. By using graphs with well-studied properties, we reduce the errors in the test specifications. For example, cyclic graphs are not acyclic. For custom graphs, we have to check this ourselves.

Quick and Nimble were added because they improve the readability of tests.

Related to: #4
See also: #1