Add docs, fix code and rename path5 to path
I renamed path5 to path because it is actually not important that the
path is on exactly 5 vertices. I chose 5 because the number is not too
small but also not too large. A too large number would make the tests
cumbersome to write. A too small number could lead to tests that do not
cover enough details.