Add reducer version of dfs (#11)
This version of depth-first search traverses the dependency graphs and
reduces the visited vertices. For example, it could be used to collect
all dependencies for a package. A few tests for this method were added.
However, I should test this in future more thoroughly.

An enum for the traverse direction was defined and a method for getting
the neighbours of a vertex.

I added additional graphs to the test package:

- path
- path with an inverted edge
- binary tree
- two disconnected components
- complement of a complete graph

Related to https://github.com/ikelax/swift-dependency-graphs/issues/1
See pull request https://github.com/ikelax/swift-dependency-graphs/pull/11