Add CI pipeline (#14)
Add CI pipeline

The pipeline runs the test on macOS and Ubuntu. It checks the formatting
and linting of the code with SwiftLint and swift-format. The spelling is
checked by Typos.

After, I have migrated all tests to swift-testing from Quick and Nimble,
I will add a job that runs the tests on Windows. Nimble uses
CoreFoundation which is not supported on Windows. swift-testing is
supported on Windows.

xcbeautify is used as reporter for failed tests. I was not able to find
out how to set the path for Linuxbrew/Homebrew such that the next runner
finds xcbeautify. So, I added two separate jobs and run the commands for
the Ubuntu job in one runner.

swift-format and SwiftLint have different default for trailing commata.
Because I prefer trailing commata, I disabled the rule from SwiftLint.

I fixed the spelling, formatting and linting errors found by the tools.

See also

For SwiftLint command
https://github.com/realm/SwiftLint/blob/main/.github/workflows/lint.yml

For xcbeautify command in GitHub Action
https://github.com/cpisciotta/xcbeautify

For Typos Action
https://github.com/crate-ci/typos/blob/master/docs/github-action.md

Swift-testing support on Windows
https://github.com/swiftlang/swift-testing

Migration to swift-testing
https://github.com/ikelax/swift-dependency-graphs/issues/15

Related to https://github.com/ikelax/swift-dependency-graphs/issues/2
See pull request https://github.com/ikelax/swift-dependency-graphs/pull/14