Add Sequence Protocol conformance
DependencyGraph conforms to the Sequence Protocol now. I added this by
adding a makeIterator() method that returns the iterator
DepthFirstSearchIterator.

DepthFirstSearchIterator iterates over the vertices in the graph in the
order a depth-first search would visit them.

The conformance to the Sequence Protocol is very useful because we get a
lot for free, for example higher-order functions like map and filter.
However, the iterator is not configurable up to now. For example, it is
not possible to adjust the traverse direction and the iterator traverses
always over the whole graph. For many applications, it is required to
only do the depth-first search in the component of the starting vertex.

I changed the vertices property in DependencyGraph to an array of V
because I did not use the dictionary so far and mostly do not see any
purpose in it. I suppose that it is usually sufficient to know which
vertices belong to the graph. Maybe, it would be better to use a Set
here.

Also, I formatted the code with swift-format.

The typo in nextVertexInDephtFirstSearch is intended since I want to
test if typos detects such errors.