Migrate tests to swift-testing (#22)
The tests were migrated to swift-testing because it is supported on
Windows unlike Nimble. Furthermore, I added a job for running the tests
on Windows to the pipeline. Since the project does not rely anymore on
Quick and Nimble for the tests, I removed the packages. swift-testing
does not have to be included as dependency.

I removed SwiftDependencyGraphsTests because the file was generated by
XCode or SwiftPM and I did not use it. I split
ContainsVertexWithTests.swift since I found the file too big and I lost
the overview.

Closes https://github.com/ikelax/swift-dependency-graphs/issues/15
See pull request https://github.com/ikelax/swift-dependency-graphs/pull/22