Implement becomesCyclicWith
I changed the signature of becomesCyclicWith because a graph cannot
become cyclic after the insertion of a vertex. For simplicity, I
implemented the method for only one edge. If more than one edge is
inserted, it has to be called multiple times. In future, I may change
this if the library is more stable and there is a significant gain in
performance.

If a graph that was acyclic has become cyclic by the insertion of an
edge, the inserted edge has to be in the cycle. Otherwise, the graph
would have been already cyclic. That's why the call of
depthFirstSearchImpl starts from tailVertex and goes in forwards
direction. If the depth-first search visits headVertex, we found a cycle
in the graph with the inserted edge. The cycle is then the found path
from tailVertex to headVertex with (headVertex, tailVertex).

I changed the vertices of DependencyGraph to a set of vertices because I
do not see any value in the dictionary. So far, I did not use any
properties of it.