Extract tested graph as property
This is done to communicate that all (or most) tests in a suite operate
on the extracted graph. If the graph is only used once, I did not
extract it.

The name of the property is the name of the graph such that it is more
clear on what kind of graph the test operates.