Add function insert(newVertex: V)
The behaviour of the function and its signature are copied from the Set
API of the Swift Standard Library. I also copied parts of the
documentation from https://developer.apple.com/documentation/swift/set/insert(_:)-nads.

Two vertices are considered to be equal iff their ids are equal.

So, if a vertex with the id of newVertex is already in the graph, the
function returns (false, vertex). Otherwise, it inserts the vertex and
returns (true, newVertex).

Related to https://github.com/ikelax/swift-dependency-graphs/issues/4