Format code
The configuration file for swift-format was moved to Tests such that it
only affects the tests file. Properties that are set by default were
removed to keep the configuration file short. I disabled the SwiftLint
rule line_length at some places because the description of a test suite
has to be on the same line as the struct declaration such that XCode
shows it while scrolling.