Insert vertex (#31)
The behaviour of the function and its signature are similar to the Set
API of the Swift Standard Library. I also copied parts of the
documentation from developer.apple.com/documentation/swift/set/insert(_:)-nads.

Similar to the Set API, we want to distinguish between inserting and
updating. This means that insert(newVertex:) does not update existing
vertices. The vertices are identified by their ID.

The configuration file for swift-format was removed because I actually
only use the default values.

Related to: https://github.com/ikelax/swift-dependency-graphs/issues/4.