Neighbours of vertex (#33)
I added a version of neighbours that computes the neighbours of a vertex
in both directions (and a few tests for both methods). neighbours was
originally added in #11.

Also, I added tests for the DFS reducer for the case that the vertex is
not in the graph because I noticed that there aren't any.

The assertion to the reduce version of depthFirstSearchImpl was added to
explain why it suffices to only check if the vertex has neighbours.

The type names were omitted to avoid repetitively and because they do
not add any value, for example express something in the domain.

The .swiftlint.yml was added to ignore the nesting rule in Tests. The
rule disallows to nest types more than 1 level deep. When using
swift-testing, nesting is often required or at least useful, since Suites
are structs. The nested configuration overwrites the main configuration.

See also: https://github.com/realm/SwiftLint#nested-configurations
See also: https://realm.github.io/SwiftLint/nesting.html
Related to: https://github.com/ikelax/swift-dependency-graphs/issues/4